Test ScanLine()
ScanLine() is a component of the tokenizer loop that
scan tokens over a strnig.
Currently, Number, Name, Operator, Bracket and Special can be correctly
scaned.
If some branch is nullable, put it at the end of branches or it will
cause dead loop.
Only use \(\) or cgroup() at top level regex and use group() otherwise.